[Ex4] Recreate Start and add correct Completed.
