--- Exercise 4/Completed/LabWorkspace.sln
@@ -1,20 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2012
-Project("{f2a71f9b-5d33-465a-a702-920d77279786}") = "LabWorkspace", "LabWorkspace\LabWorkspace.fsproj", "{D863E586-4A83-4807-8721-2C3DE0E8CA8D}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D863E586-4A83-4807-8721-2C3DE0E8CA8D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D863E586-4A83-4807-8721-2C3DE0E8CA8D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D863E586-4A83-4807-8721-2C3DE0E8CA8D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D863E586-4A83-4807-8721-2C3DE0E8CA8D}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(MonoDevelopProperties) = preSolution
-		StartupItem = LabWorkspace\LabWorkspace.fsproj
-	EndGlobalSection
-EndGlobal
