--- Exercise 4/Completed/LabWorkspace/AssemblyInfo.fs
@@ -1,18 +1,0 @@
-﻿module LabWorkspace.AssemblyInfo
-
-open System.Reflection
-open System.Runtime.CompilerServices
-
-[<AssemblyTitle("LabWorkspace")>]
-[<AssemblyDescription("")>]
-[<AssemblyConfiguration("")>]
-[<AssemblyCompany("Xamarin")>]
-[<AssemblyProduct("")>]
-[<AssemblyCopyright("Xamarin")>]
-[<AssemblyTrademark("")>]
-[<AssemblyVersion("1.0.0.0")>]
-()
-// The assembly version has the format {Major}.{Minor}.{Build}.{Revision}
-//[<assembly: AssemblyDelaySign(false)>]
-//[<assembly: AssemblyKeyFile("")>]
-
