--- Exercise 4/Completed/LabWorkspace/Component1.fs
@@ -1,5 +1,0 @@
-﻿namespace LabWorkspace
-
-type Class1() = 
-    member this.X = "F#"
-
