--- Exercise 4/Completed/LabWorkspace/Exercise2.fsx
@@ -1,8 +1,0 @@
-﻿let generateTriangleNumbers (state, counter) = 
-    Some(state, (state + counter+1, counter+1))
-
-let triangulars = Seq.unfold generateTriangleNumbers (1,1)
-
-triangulars 
-    |> Seq.take 4
-    |> Seq.toList
