--- Exercise 4/Completed/LabWorkspace/Exercise4.fsx
@@ -1,4 +1,0 @@
-﻿type Registrant = 
-    | Attendee of email:string
-    | Speaker of room:int * title:string
-    | Volunteer
