+++ Exercise 4/Completed/Registration/Registration/Registrant.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace Registration
+{
+    public abstract class Registrant
+    {
+    }
+}
