+++ Exercise 4/Completed/Registration/Registration/Volunteer.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace Registration
+{
+    public class Volunteer : Registrant
+    {
+        public Volunteer()
+        {
+        }
+    }
+}
