+++ Exercise 4/Completed/Registration/RegistrationFSharp/Script.fsx
@@ -1,0 +1,8 @@
+﻿// Learn more about F# at http://fsharp.org. See the 'F# Tutorial' project
+// for more guidance on F# programming.
+
+#load "Registration.fs"
+open RegistrationFSharp
+
+// Define your library scripting code here
+
