+++ Exercise 4/Completed/Registration/RegistrationFSharp/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="System.ValueTuple" version="4.3.1" targetFramework="net461" />
+</packages>
