--- Exercise 4/Start/Registration.sln
@@ -1,20 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 11.00
-# Visual Studio 2010
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Registration", "Registration\Registration.csproj", "{BD6FF681-48C9-4091-A181-09D1103D91AC}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{BD6FF681-48C9-4091-A181-09D1103D91AC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BD6FF681-48C9-4091-A181-09D1103D91AC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{BD6FF681-48C9-4091-A181-09D1103D91AC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{BD6FF681-48C9-4091-A181-09D1103D91AC}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(MonoDevelopProperties) = preSolution
-		StartupItem = Registration\Registration.csproj
-	EndGlobalSection
-EndGlobal
