--- Exercise 4/Start/Registration/Attendee.cs
@@ -1,14 +1,0 @@
-﻿using System;
-
-namespace Registration
-{
-	public class Attendee : Registrant
-	{
-		public Attendee (string email)
-		{
-			Email = email;
-		}
-	
-		public string Email { get; private set; }
-	}
-}
