--- Exercise 4/Start/Registration/Registrant.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace Registration
-{
-	public abstract class Registrant
-	{
-	}
-}
