+++ Exercise 4/Start/Registration/Registration/Attendee.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace Registration
+{
+    public class Attendee : Registrant
+    {
+        public Attendee(string email)
+        {
+            Email = email;
+        }
+
+        public string Email { get; private set; }
+    }
+}
