+++ Exercise 4/Start/Registration/Registration/Volunteer.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace Registration
+{
+    public class Volunteer : Registrant
+    {
+        public Volunteer()
+        {
+        }
+    }
+}
