--- Exercise 4/Start/Registration/Speaker.cs
@@ -1,16 +1,0 @@
-﻿using System;
-
-namespace Registration
-{
-	public class Speaker : Registrant
-	{
-		public Speaker (string title, int room)
-		{
-			Title = title;
-			Room = room;
-		}
-
-		public string Title { get; private set; }
-		public int Room { get; private set; }
-	}
-}
