--- Exercise 4/Start/Registration/Volunteer.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace Registration
-{
-	public class Volunteer : Registrant
-	{
-		public Volunteer ()
-		{
-		}
-	}
-}
-
